backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
pandas>=2.2.0
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.4
requests==2.31.0
pydantic>=2.7.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Multi-Agent Framework (Python 3.9 compatible)
langgraph==0.2.28
langchain==0.2.16
langchain-google-genai==1.0.10
langchain-core==0.2.39
langchain-community==0.2.16

# Financial Data
yfinance==0.2.18
alpha-vantage==2.3.1

# Additional utilities
python-dotenv==1.0.0
aiofiles==23.2.1